+++ CHANGELOG.md
@@ -13,7 +13,17 @@
   Change extension configuration `oidcClientScopeSeparator = ,` for old behaviour.
 - Feature: Added extension setting `enablePasswordCredentials` to disable password-authentication.
 - Method `getFreshAccessToken()` now actually returns the fresh access token.
+- Breaking: Configure user claim mapping without TypoScript.
 
+  Mapping is now done via a YAML file: `config/system/oidc.yaml`
+  ```
+  providers:
+    default:
+      mapping:
+        fe_users:
+          name: '<name> // <family_name> <given_name> // <email>'
+          ...
+  ```
 
 ## Version 4.0.0